+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## v1.x.x
+## v1.4.0
 
 <!--Releasenotes start-->
-- Added a shuffle button to the extension popup. This button will shuffle from the most recently visited channel and will be visible when you are on a YouTube page.
+- Added a shuffle button to the extension popup that will shuffle from your most recently visited channel. When shuffling using this button, a new tab will open that must remain open while the shuffle is being prepared.
+- When an error is encountered, the full error message and a hint for what to do about it will be displayed.
 - Some updates to the internal handling of API keys.
+- Improved maintainability of the codebase.
 <!--Releasenotes end-->